Re-enable length check, use bytes constructor
We're using a forceful unwrap on the failable constructor to trap in the case that the length check fails (since the publicKey we get comes straight from Sodium, so if that fails something has gone very wrong. Though in general use there isn't a good reason to expose the nil case since the reason for failure would be that Sodium somehow malfunctioned).