Add manual decompression logic
Apple seeimingly broke the x963Representation constructor when releasing
the new compressedRepresentaion constructor, the former no longer accepts
compressed points which is a BC break. This would be fine, except the new
method cannot be used by newer OS versions if we want to support older OS
versions at the same time. Either we need to avoid using the x963
constructor, or only support macOS 13 and similar. This is pretty new,
and not even GitHub supports it yet for pipelines. It's not ideal, but
it seems the only workable alternative is the implement handling of
compressed points manually. This is annoted pretty heavily to follow the
steps and notation in the spec.